+++ test/parsers/htmlParser.js
@@ -1,5 +1,5 @@
 require('chai').should();
-const HtmlParser = require('../../lib/parsers/htmlParser');
+const parseHTML = require('../../lib/parsers/htmlParser');
 const YAML = require('yaml');
 const fs = require('fs');
 const path = require('path');
@@ -11,7 +11,7 @@ describe('HtmlParser', function () {
       const file = fs.readFileSync(path.resolve(__dirname, CONFIG_FILE), 'utf8');
       const yamlData = YAML.parse(file);
 
-      this.output = HtmlParser.parse(yamlData.paths['/movies/{actor}']['get'].responses, '', '');
+      this.output = parseHTML(yamlData.paths['/movies/{actor}']['get'].responses, '', '');
     });
 
     describe('#functionality()', function () {