+++ test/parsers/routeParser.js
@@ -1,9 +1,8 @@
 require('chai').should();
 
-const YAML = require('yaml');
 const fs = require('fs');
 const path = require('path');
-const RouteParser = require('../../lib/parsers/routeParser');
+const parseRoute = require('../../lib/parsers/routeParser');
 
 const CONFIG_FILE = '../resources/config_test_example.yaml';
 
@@ -12,7 +11,7 @@ describe('RouteParser', function () {
     before(function () {
       const file = fs.readFileSync(path.resolve(__dirname, CONFIG_FILE), 'utf8');
 
-      this.output = RouteParser.parse('/movies/{actor}', 'get');
+      this.output = parseRoute('/movies/{actor}', 'get');
     });
 
     describe('#functionality()', function () {